+++ web/source/kmwembedded.ts
@@ -40,7 +40,7 @@ namespace com.keyman.osk {
     }
   };
 
-  // iOS now relies upon native-mode popup key management, so we only implement these hybrid-targetted
+  // iOS now relies upon native-mode popup key management, so we only implement these hybrid-targeted
   // methods when embedding in Android.
   let device = com.keyman.singleton.util.device;
 
@@ -56,10 +56,8 @@ namespace com.keyman.osk {
       if(key['subKeys'] && (typeof(window['oskCreatePopup']) == 'function')) {
         var xBase = dom.Utils.getAbsoluteX(key) - dom.Utils.getAbsoluteX(this.kbdDiv) + key.offsetWidth/2,
             yBase = dom.Utils.getAbsoluteY(key);
-        
-        if(util.device.formFactor == 'phone') {
-          this.prependBaseKey(key);
-        }
+
+        // #3718: No longer prepend base key to subkey array
 
         this.popupBaseKey = key;
         this.popupPending=true;
@@ -317,15 +315,6 @@ namespace com.keyman.text {
     keymanweb.modelManager.register(model);
   };
 
-  keymanweb['showNewSuggestions'] = function() {
-    let keyman = com.keyman.singleton;
-
-    if(keyman['osk'].banner['activeBanner'] instanceof com.keyman.osk.SuggestionBanner) {
-      let banner = keyman['osk'].banner['activeBanner'];
-      banner.rotateSuggestions();
-    }
-  }
-
   /**
    * Function called by Android and iOS when a device-implemented keyboard popup is displayed or hidden
    * 
@@ -493,8 +482,9 @@ namespace com.keyman.text {
    *  @param  {number}  shift  shift state (0x01=left ctrl 0x02=right ctrl 0x04=left alt 0x08=right alt
    *                                        0x10=shift 0x20=ctrl 0x40=alt)
    *  @param  {number}  lstates lock state (0x0200=no caps 0x0400=num 0x0800=no num 0x1000=scroll 0x2000=no scroll locks)
+   *  @return {boolean} false when KMW _has_ fully handled the event and true when not.
    **/            
-  keymanweb['executeHardwareKeystroke'] = function(code, shift, lstates = 0) {
+  keymanweb['executeHardwareKeystroke'] = function(code, shift, lstates = 0): boolean {
     let keyman = com.keyman.singleton;
     if(!keyman.core.activeKeyboard || code == 0) {
       return false;
@@ -525,7 +515,9 @@ namespace com.keyman.text {
     try {
       // Now that we've manually constructed a proper keystroke-sourced KeyEvent, pass control
       // off to the processor for its actual execution.
-      return keyman.core.processKeyEvent(Lkc) != null;
+
+      // Should return 'false' when KMW _has_ fully handled the event and 'true' when not.
+      return !keyman.core.processKeyEvent(Lkc);
     } catch (err) {
       console.error(err.message, err);
       return false;