+++ windows/src/developer/TIKE/project/Keyman.Developer.UI.Project.UfrmNewProjectParameters.pas
@@ -179,6 +179,7 @@ begin
   inherited;
   editPath.Text := FKeymanDeveloperOptions.DefaultProjectPath;
 
+  dlgBrowse := TBrowse4Folder.Create(Self);
   dlgBrowse.InitialDir := editPath.Text;
   dlgBrowse.Options := [OnlySelectFileSysDir, ShowEditBox, UseNewDialogStyle];
   dlgBrowse.Root := Desktop;