+++ HISTORY.md
@@ -1,5 +1,32 @@
 # Keyman Version History
 
+## 14.0.164 alpha 2020-10-20
+
+* feat(common/models):  context tracking of accepted Suggestions (#3663)
+* feat(common/models): context reversion modeling (#3685)
+* fix(windows): RefreshKeyboards loses some profiles (#3714)
+* fix(windows): icons missing in Configuration (#3717)
+* fix(windows): tweak scrolling in keyboard menu (#3719)
+* fix(developer): crash creating basic project (#3716)
+
+## 14.0.163 alpha 2020-10-19
+
+* fix(web): Remove base key from popup keys (#3718)
+
+## 14.0.162 alpha 2020-10-18
+
+* fix(web): Let embedded device handle K_TAB or K_ENTER (#3664)
+
+## 14.0.161 alpha 2020-10-16
+
+* fix(common/core/web): fixes revert event bug (#3709)
+
+## 14.0.160 alpha 2020-10-14
+
+* feat(common/models): disables "keep" when word is not suggestion otherwise (#3700)
+* feat(common/core/web): selective wordbreak swallowing after accepting suggestions (#3702)
+* refactor(common/models): extract Outcome type (#3705)
+
 ## 14.0.159 alpha 2020-10-12
 
 * fix(web): unit test script failure on compilation failures (#3597)